Updating scraper
i-league.org is a dead end now that they are loading their content in a
iframe from somewhere else. So I am using goal.com to scrape the
standings table. They seem to be behind by a few matches but that is a
compromise we're going to have to make IMO till we find a better way